refactor: remove some deprecated Bonus classes
